potentially solved db locking bug
aug-db was managing to get into a state where the sqlite
db would be permenantly locked, with no other process
able to write to the db. turns out that when in
'auto-commit' mode (default when BEGIN ... COMMIT is
not used), the db will not give up a write lock
until concurrent reads (select) are finalized even if the
write (i.e. update, insert) has finished. when the user
selects a result item the chosen_at column is updated
in the corresponding blob row. however this always
coincides with a select which is populating the
new result set. thus the update needs to be wrapped in
a BEGIN ... COMMIT so that sqlite will commit that
transaction without waiting for the select to finish.

initial tests show that this change has solved the
db locking issue.