working on factoring out query stuff from ui_state
this needs to be done to ease changing the design in a way such
that we are not holding sqlite statements open all the time.
keeping a sqlite statement open causes the shared db lock to
to stay allocated and prevents any other processes from being
able to acquire write locks.